refactor: 컴포넌트 폴더 정리
- RecipeWidget을 presentation/component로 이동
- Recipe 위치 이동
