+++ lib/data/model/recipe.dart
@@ -1,0 +1,6 @@
+class Recipe {
+  final String name;
+  final String imageUrl;
+
+  const Recipe({required this.name, required this.imageUrl});
+}
